Refactor TabSqlQuery into its own TS component and implement ABAC for SQL editor create, update, delete
